Remove --Xuse-k1-uast from the metalava command line
This will use the new and improved Kotlin parser (K2). Among other
things it changes the project structure to better support the Kotlin
multi-platform code. While there is very little actual Kotlin code
consumed by Metalava in the Android build and none of it contributes to
the API surface, both Java and Kotlin are processed at the same time
and through the same project so K2 will have some impact on it.

AndroidX is already ~90% switched over to using K2 and updating Android
to match will ensure stability and consistency moving forward and
simplify Metalava in the medium to long term as it no longer has to
support both K1 (the current parser) and K2.

Bug: 326652553
Test: ./gradlew
Change-Id: Ie3416d7d33aa13e5eb080cbb557ce6712a62935e