Make more dex optimisation flags configurable with select(), in particular optimize.enabled (reland).
This is necessary to be able to ramp them with build flags.

This CL should not otherwise change the logic how the effective values
for the optimisation flags are determined (from defaults and other flag
values).

Add ShrinkByDefault to accommodate for it to be true by default for
android_app.

Also extend some flag docs to describe the logic more accurately.

Test: m
Test: m EMMA_INSTRUMENT=true EMMA_INSTRUMENT_FRAMEWORK=true
Test: cd build/soong/ ; go test ./...
Flag: EXEMPT UNSUPPORTED_LANGUAGE
Bug: 349177664
Bug: 395120653
Change-Id: I9163afb2a6c0dab08a23aee84e24917e134f62fb