Add a non-select()'able property force_disabled to disable R8 optimisation.
To allow select() on the ordinary optimize.enabled property, it has to
statically add some SDK dependencies, regardless whether R8 is enabled
or not by the select(). That's due to the evaluation order in Soong.

Those SDK dependencies lead to cycles between SDK libraries, and this
flag is used to address that. In particular it disables R8 for stubs,
but it can also be used explicitly in .bp files.

R8 is also disabled by default in coverage builds, since they force
source stubs to be used (cf. b/302320354) which leads to cycles. This
can have the effect that non-SDK libs with optimize.enabled=true may
get disabled too. In such cases it can be overridden with
optimize.force_disabled=false.

Test: m
Test: m EMMA_INSTRUMENT=true EMMA_INSTRUMENT_FRAMEWORK=true
Test: cd build/soong/ ; go test ./...
Flag: EXEMPT UNSUPPORTED_LANGUAGE
Bug: 395120653
Change-Id: Iecd42480b5910c2f77966ec01e6c880e49b439af