Add support for R8 on SDK libraries in the boot classpath without Soong dependency cycles.
To allow select() on the ordinary optimize.enabled property, it has to
statically add some SDK dependencies, regardless whether R8 is enabled
or not by the select(). That's due to the evaluation order in Soong.

Those SDK dependencies lead to cycles between SDK libraries when source
stubs are used, since they will depend on each other's implementation
libs then. (Source stubs are typically only used in SDK and coverage
builds, not in regular builds for devices.) This CL adds a flag
`force_disabled_with_source_stubs` to address that.

Also disable R8 when building the stubs for SDK libraries, regardless of
other flags.

This is a partial reland of commit
7ac158ac7e2db4d197fdfaeeee09f439d6dca01f, and it fixes b/451577887 by
force disabling R8 whenever source stubs are used for any reason, in
particular in SDK builds.

Test: Build sdk and mainline_modules_sdks in heavy presubmit
Test: m EMMA_INSTRUMENT=true EMMA_INSTRUMENT_FRAMEWORK=true
      and verify that jars don't regress on size
Test: Enable optimisation by default on the boot classpath and verify
      that R8 doesn't get disabled in a build for a device.
Test: cd build/soong/ ; go test ./...
Test: m WITH_DEXPREOPT=false droid
      and check that there are no significant differences in
      $ANDROID_PRODUCT_OUT with and without this topic.
Flag: EXEMPT UNSUPPORTED_LANGUAGE
Bug: 395120653
Change-Id: Idd2aecd529ad01bb64d242d68b7a9f9f502cf0d8