Add SOONG_GEN_COMPDB_MODULES to filter compdb generation
This change introduces the `SOONG_GEN_COMPDB_MODULES` environment variable. When set, it allows specifying a space-separated list of module names to either include or exclude from the generated `compile_commands.json`. If the value starts with a "-", the listed modules are excluded; otherwise, only the listed modules are included.

Bug: 442496243
Test: manual

Change-Id: Ia9ba814c314d1934fb90bda4b8cf43166916f641