Add a UUID to builds
Add a unique identifier to builds to allow correlating logs from a build
with the exact build that they came from.  The UUID is computed based
on a hash of the BUILD_TARGET_NAME and BUILD_ID environment variables.
The value is placed in the ro.build.uuid and ro.{partition}.build.uuid
properties, and copied to $DIST_DIR/BUILD_UUID for indexing.

Bug: 439900398
Test: manual
Flag: EXEMPT HOST_MODE
Change-Id: Idf03300a4fc2851508078d4ec008580cb5b1858b