Use BUILD_SYSTEM_FINGERPRINT for ro.system.build.fingerprint
To remove any product-specific information from the system image,
define a fingerprint for the system image without the product name.

We have thumbprint that is a fingerprint without the device name, but
it has a different format from the fingerprint. So we defined a new
fingerprint only for the system image.
The system fingerprint uses PRODUCT_SYSTEM_NAME for its device name.

Bug: 437803910
Test: getprop ro.system.build.fingerprint
Test: vendor/google/tools/flashall
Change-Id: I9028bfca32f016b3c64c6bc27b98e1a3e4fd80f1