Include module-info independently of androidmk
Previously, module-info information was tied to if the module
implemented an androidmk function. (and in some cases, what that
function returned) Change it to always output the module-info for that
module as long as SkipAndroidMkProcessing is false.

Soong-only builds were already doing this, so this makes soong-only
and soong+make builds' module-info files identical. (except for 1 module
that only exists in soong-only builds)

This does cause a lot more modules to be added to the module-info.json
file.

Bug: 444756259
Test: Manually
Change-Id: I0cdb96b8b609d750cff5d9ffd26647c65f7f86e2