Move module-info build to a parallel singleton
Currently, the soong module-info file is built as part of the androidmk
singleton, which isn't ideal because that singleton is one of the few
non-parallel singletons. Move it to its own parallel singleton to help
with analysis time a little.

This also makes the code more consistent between soong-only and
soong+make builds.

Bug: 444756259
Test: Diff'd module-info.json before/after on both soong-only and soong+make builds, saw no difference
Change-Id: I0cc99025724ae7547f01fb90f3955915dd07a596