Remove valid filesystem module checking for custom partition entries
Instead, this change sets the `custom_partitions` property of the
`android_device` module and the `chained_partitions` property of the
vbmeta image module purely based on the board variables to match the
behavior in make.

Note that this will lead to panic when building soong-only in products
that depend on non-filesystem custom partition entries.

Test: go test ./...
Bug: 445246294
Change-Id: I7d088351f00db9cc497ecb5bbbc41471c7b9d4f9