Copy vendor-bootconfig.img to staging dir in Soong only build
Contrary to its name, vendor-bootconfig.img is not a partition image
file but is a txt file that contains the bootconfig parameters. The
support to generate this in Soong was added in aosp/3367784, but the
support to copy this to PRODUCT_OUT was never added.

This change adds such support by copying the vendor boot image-provided
bootconfig to the staging dir in the android_device module. This is only
done when the vendor_boot_partition_name property is specified, given
that the copying is only done when vendor boot image is built in Make.

Test: aninja -t path droid out/target/product/vsoc_x86_64_only/vendor-bootconfig.img
Bug: 444833782
Change-Id: I9f2667858464d7e9a86fac6d29d796d58999cc92