find_input_delta outputs content additions and deletions.
Prior to this change, if the --inspect flag was passed,
find_input_delta would output changed contents changed zips/jars
but if a zip/jar was added or removed, it would simply output the file
as a whole being added or removed. The added/removed contents of it
was absent.

The incremental kotlin client doesn't look at the zip/jar directly. It
wants to know what source files have actually been added or removed.

With this change, when a zip/jar is added or removed, its entire
contents is correpondingly listed as added or removed.

Bug: 447346986
Test: m TestApisReflection
Change-Id: I2ae084106b2cf129e2a6ccbb918eb70bb901856d