Inspect the contents of src jars for changes.
Prior to this change, if source jars contents changed, the client
would not pick up on the changes unless other code in the target
referenced the changed code.

While this is usually true, there are cases where an annotation
processor can change its behavior without the related source code
actually changing. This manifested in the associated bug where new
code was being generated by an annotation processor that the target
source code did not directly reference. Because the compiler was not
told of these new, generated source files, it would fail to compile.

With this change, we now pass the --inspect argument to
find_input_delta. We also add a new, -src-jars-dir argument to where
the unzipped contents of the src jar already existed.

We do not inspect the srcJarDir directly because it is deleted and
recreated with every invocation. It would appear to the client as if
the files are changed with every run. Using the --inspect option, we
can rely on the hash that is generated as part of the zip process to
detect changes.

Testing this involved checking out a previously green build, running
`m TestApiReflections`, syncing forward to a known bad build, and
then seeing that this cl resolved the issue.

Fixes: 44734698
Test: m TestApiReflections
Change-Id: If69fab095f28ebe20e18226e4bfc881550e58560