Add SOONG_ACTION_SANDBOXING to toggle sandboxing
Add an environment variable that can be set to cause ninja to sandbox
all the actions in the build. This will allow us to start working
on fixing the missing dependencies.

I'm only applying this to the main build phase, not the bootstrap
build phase for now, as that's the phase that has the most effect, and
the soong_build action we know has lots of undeclared inputs essentially
by design. (Like implicitly finding AndroidManifest.xml files for
android modules)

Bug: 448484009
Test: Manually
Change-Id: I9292811060d8f5ff5578fdae5bf2027fff230a00