rust: Add implicit _nostd elision for transition
In order to avoid boiling the ocean with the `no_std` variant
transitions in our 3p crates repository, when resolving Rust
dependencies we will try truncating off the `_nostd` suffix on a module
name if it doesn't exist.

We should likely try to clean this up later so that new code does not
rely on it.

Bug: 446970402
Test: Build Virtualization after making no_std serde variant
Test: New Soong test
Change-Id: I6502ea22e35026eb6ab6b88431e5a660d5fc49c4