rust: Add no_std_rlibs dep variety
Similar to `rlibs` or `dylibs`, this allows modules to forcibly select a
`no_std` variant of a dependency even if they are not `no_std`
themselves.

The primary reason a user would want to do this is to have their
`std`-linked module use the same dependency as a transitive dependency
that goes through a library that only has a `no_std` variant.

Another reason would be to create a test module which links a `no_std`
variant which might have a different implementation.

There are also a number of legacy configurations that may be in error
that this allows us to leave identical as we migrate.

Bug: 446970402
Test: Add `no_std_rlibs` to `Virtualization` for `_nostd` from std mods
Test: New Soong test
Change-Id: Ic3008b17ff079b22fa7443c9744fb8f79c7dd76a