rust: Reorder linkage splits
When running a mutator, after `split` returns, Soong creates each module
variant with a dependency edge to the default variant of the module.
That dependency edge is removed after the `Mutate` step completes.

The problem that arises is that in the case where:

1. libfoo-dylib depends on libstd
2. libstd depends on libc
3. libc depends on libfoo-nostd

If dependencies for `libfoo-dylib` are processed before dependencies for
`libfoo-nostd`, there is a temporary circular dependency. Soong will
detect this and bail out.

Since `rlib-core` components can be used in the two others, but
`rlib-std` and `dylib-std` cannot be mixed, the only real requirement is
that `rlib-core` must come first, so we now have the right order.

Bug: 406843571
Test: Link no_std icu4x into libc without getting a cycle
Test: New soong test
Change-Id: Ic2c2e5d1fd675f1c3ed5b76560c8ce82b084d1e6