Treat ErrorProne as a plugin.
The current ErrorProneClasspath hack hardcodes the fact that
ErrorProne is a prebuilt JAR. Replace this with an explicit
dependency on error_prone_plugin, a new java_plugin module
defined in external/error_prone. This will allow us to build it
from source.

The trade-off is that all dependencies of ErrorProne will need
to have ErrorProne checking explicitly disabled to avoid circular
dependencies.

Bug: 448761419
Test: presubmit
Flag: EXEMPT HOST_MODE
Change-Id: Ib396c2f6b8601594acbd85628542cd74b6a0574b