Add an analyzer script for listing hidden dependencies
Currently, it is difficult to identify and track all hidden dependencies
in a single build, which are needed to enable action sandboxing for all
ninja targets.

This change introduces an analysis tool to address this problem. The
script utilizes the ninja log generated by the Ninjago of ABFS and the
local cache database of the ABFS to find the hidden dependencies of all
targets that are built.

The scripts operates in two modes:
1. `analyze`: Parses a `.ninja_log` file or its hash, runs `abfs inspect`
   concurrently on each action's output to find hidden dependencies, and
   saves the results into sharded protobuf files.
2. `convert`: Converts the generated binary protobuf files into a
   human-readable textproto format for inspection and debugging.

Test: Run ABFS build, manually test the script
Bug: 447462375
Change-Id: I1f745c041e7aa8c782076abac4ab85bc7fe2597a