Use new --output flag of systemfeatures-gen-tool
The Java runtime currently has a bug where sometimes it will output
a warning to stdout when running a java/kotlin tool. Since
systemfeatures-gen-tool produces output to stdout, that output gets
mixed with the warning and causes the build to fail. Write directly
to a file instead as a workaround.

Bug: 449768258
Test: m systemfeatures-gen-srcs
Change-Id: If416d198bac6490f2f3e530e8566329dad513aac