Tighten directories passed to javac --patch_module
Using --patch_module=java.base=.:out/soong results in javac searching
for java.base sources in the root of the source tree.  If sources were
accidentally created there, for example by unzipping a srcjar during
debugging, that can cause javac to incorrectly use those source files
instead of jars on the classpath.

Tighten the list of paths passed to --patch_module to be the directories
that contain the source files, the classpath, and the module output
directory to cover the unzipped srcjars.

Bug: 436225239
Test: TestPatchModule
Test: build all modules that have patch_module set.
Change-Id: Ie3186842a0ec16fcccd03978c687daae8c6ce879