Queue RBE shutdown output for later
Shutting down RBE prints the RBE stats, which replaces the last
status line from ninja that tells the user what the build just
finished.  Store the output for later, and print it after the
status output has shut down and set the cursor to a location that
won't overwrite the status.

Test: manual
Change-Id: I92bf3e8564ce0e122b452753747463f788a2c07a