Switch depmod to use zip files as inputs
This is the first step in being able to accept arbitrary zip files full
of modules from CIPD, rather than only a list of files from the source
tree.

We also change the `{.modules}` output to a `{.modules.zip}`, and
support taking that zip file in via `system_dep`. The single user in
fsgen has been updated.

Bug: 452027180
Test: compare installed lib/modules folders before/after
Change-Id: I1dc68e205e64e63bd3a1f3f116009381f6536fef