rust: Add Force_rust_staticlib for CC modules
Normally we produce a single Rust staticlib containing all Rust static
dependencies only in the final link to avoid symbol collisions. This
works for all things that are built in Soong. However, in some cases
where the static library is used outside of Soong, such as in the NDK,
we need to force inclusion of the Rust staticlib to ensure there are no
missing symbols.

This adds force_rust_staticlib to force generation and inclusion of the
Rust staticlib for these cases.

Bug: 454363909
Test: m libc.ndk_static
Change-Id: I6835d0a74dab8526f355da5d70993d4312624247