Add API to disable sandboxing for build rules
This change introduces an API to mark rules as sandbox-incompatible.
Static rules can set SandboxDisabled: true in their RuleParams, and
dynamic rules can use the RuleBuilder.SandboxDisabled() method.

When this API is used, Soong emits a sandbox_disabled = true variable
for the rule in the generated ninja file. This variable instructs a
modified Ninja to bypass the sandbox wrapper for that specific action.

Test: manual
Bug: 454133382
Change-Id: I78040d771b465d4d4bee8d94b09ea89ad5599f07