Split "module with bad pkg_path format" python test to avoid race condition
The "module with bad pkg_path format" test contained two modules
with errors and asserted that both errors were reported, but
processing new modules stops as soon as the first error is
encountered.  If the other module was already being processed the
second error will also be reported, but if not the second error
will be missing.  Split the test into two that each handle
a separate error case.

Bug: 430344528
Test: ABTD
Change-Id: I2c5421393796757e0b3e1897fe179e0a10d3d109