Temporarily disable Sbox() panic check for SandboxDisabled()
Dynamic rules that set Sbox() should ideally be able to be action
sandboxed already, but there are some inputs that are not yet specified
(e.g. rsp files).

This change temporarily disables the enforcement that Sbox() and
SandboxDisabled() cannot be set at the same time to unblock building
with `SOONG_ACTION_SANDBOXING` env var.

Test: m nothing
Bug: 454133382
Change-Id: I9740b9cb83e70be240b004bd612177291ad188b1