rust: Trigger traditional missing module error when searching variants
Previously, we'd emit an error that indicated:
1. What module you were looking for
2. What linkage you were asking for
3. What stdlib linkage were allowed

However, despite containing all the information, this error message was
more confusing for a number of users compared to the usual Soong error
message.

In the event that we were going to emit the error message before, we
instead intentionally add a missing dependency to trigger the regular
Soong error message.

Bug: 452749777
Test: TH
Test: Intentionally misspell Rust module dep, observe error
Change-Id: I65150dd54e9390bca2ec945a538fb2ee533679ad