Use module stems for static libraries.
The new libc.ndk_static library which is the full libc.a (the normal
libc module does not include the Rust components) is not being
installed correctly because apparently install stems were never
implemented for static libraries. This does that, which fixes the
install name for the NDK's libc.a.

Test: m out/soong/ndk.timestamp && find out/soong/ndk/sysroot \
    -name libc.a
Bug: https://b/454363909
Change-Id: I5cee23e7ccdc1a00b15596e0546b6fe09d952981