Split llndk_libraries_txt into a Singleton and a Module
llndk_libraries_txt is implemented as a SingletonModule in order to
make output files of the Singleton available to Android.bp files
as the output file of a Module.  Split it into a Singleton and a
Module by using a OnceKey to agree on an output path for the Singleton.

Bug: 456503585
Test: compare all Soong output files
Change-Id: Ic56b30b154d748546c913866411a67ca0eb739ef