rust: Don't copy filegroup files into out/
Filegroup files were incorrectly getting copied into the gen directory
when included in `srcs`. This is unnecessary as adding filegroups to
`srcs` should just be used for indicating that these inputs are required
for RBE.

We still want the outputs of source producers to be copied to the gen
directory, so add a field to distinguish one from the other.

Bug: 407072297
Test: filegroup files no longer copied to `out/` directory for module
Change-Id: Icb9beefb71d68cb572f6d9e7b47f45e7d23d22c4