Add RELEASE_REDACT_FLAGS_WITHOUT_OVERRIDE
When true, the on-device {partition}/etc/build_flags.json files do not
contain any flags which get their final (and empty) value from the flag
declaration.  If the fallback value is neither boolean false nor the
empty string, it will be present.

This largely mirrors how read-only aconfig flags are not present in the
on-device artifacts when they are disabled.

Bug: b/459554451
Test: manual, TH
Change-Id: Ia5afe94fed1ea72c395ad81e77c6d9bc61f3c6e8