Fix stale class files lingering after partial compile.
Currently if a A.java generates A.class and A$1.class,
and if a change to A.java now limits this to only
A.class, A$1.class still lingers in output jar, during
partial compile.

To fix this, we mandatorily remove every class file
generated out of java files in compilation set.

Bug: 458008991
Test: build/soong/tests/java_partial_compile_test.sh
Change-Id: I79d9d23da9afca25351a9cecd61bd5bc40dbcda9