Recompile everything during partial compile when tools change
If a previous run of generate_incremental_input produced a bad list
of files to recompile then even fixing generate_incremental_input
won't result in the correct outputs until the source file that needed
recompiling is modified again.  Add the path to generate_incremental_input
and dependency_mapper to the tools parameter of findInputDelta so that
all the inputs get recompiled when the tools are changed.

Bug: 458008991
Test: manual
Change-Id: I01b6e9d7d8da8a9a5c7dc4100aa4365a3d8c6976