Introduce check_api.current.default_in_droid property
This change reverts ag/36837363 and partially reverts aosp/3467665.

ag/36837363 is incompatible with the recent changes to update-api, which
now runs after the main ninja execution is run. Given that update api is
now always run after checkapi, checkapi should not be a transitive
dependency/validation of update-api.

The newly introduced default_in_droid property allows the checkapi
timestamp to be added as a dependency of droid when set to true. This is
automatically set to true for sdk_library modules that do not utilize
from-text stubs.

Test: m nothing
Test: aninja -t path droid out/soong/.intermediates/libcore/mmodules/intracoreapi/art.module.intra.core.api.stubs.source/android_common/everything/check_current_api.timestamp
Bug: 458204719
Change-Id: Ic85c910fe90cfafb352f98f18a1efe91fa805a14