VariableMutator: Avoid GC churn due to productVariables copying
reflect.ValueOf(x) copies the value if the passed interface{} is the
struct itself. When passed a pointer, reflect operates on the reference
instead, avoiding the copy.

This line was found consuming significant time calling into
runtime.convT. Initially, this looked like a red herring, but upon
closer look, it become clear that this function runs on modules with
defaulted product configs as well and was not hitting the
setVariableProperties path often. Therefore the overhead really was in
the ValueOf instantiation, and copying seems to eliminate that from perf
profiles.

See http://shortn/_zeuMnGXjqN for a profile, where the overhead in
VariableMutator is completedly eliminated. Note that the profile
contains other random wins / losses but I suspect this is run-to-run
variance.

Bug: None
Test: SOONG_CPU_PROFILE=profile m nothing --no-skip-soong-tests
Change-Id: If428e5a6999a03c3ef47376d475f95d69dd71c74