Silence ABI mismatch error by adding suggested flag
An ABI mismatch occurs in the 1.91 update surfacing preexisting broken behavior. We are turning it off to pursue the update and have created follow-ups for future work. Various Rust libraries (core, alloc,..) are built with sanitizer turned off. Follow-up work b/460816174  will explore turning it on. Putting the flag in sanitizer.go is not sufficient as another library  `generated_rust_staticlib`, gets pulled in by bionic. We moved the suppression flag to builder.go so all the right libraries get the flag and supreses the ABI mismatch error.  We have a follow-up ticket to fix the underlying warning triggered by bionic (b/460816174)

bug: b/454030311, b/460816174, b/458140256
Test: m rust
Change-Id: I7b27e6137b27fbab33afcc0432c9cf42083b3fea