Make vendor/product cflags configurable
Usually making cflags configurable is enough but we
found ourselves in need to speficially target shared
code that is built for both system and vendor.

In our case we want device flags to only affect the
vendor build, while leaving the standard target unaffected.

The same logic is used for product modules, so it gets the
same functionality as a side effect.

Test: m nothing --no-skip-tests
Change-Id: I0c0d5edc5a85e5caede1c36c0462094a3cbdb1b6