Support extra zip files in packaging, use in apex
We have this support in make-mode and special-cased in the apex logic,
but the soong-only packaging did not support it. Add support for this in
packaging specs, and use that in the apex logic rather than special
casing appSets.

I'm planning on using this for kernel modules, which will need support
in both soong packaging and apex packaging.

Bug: 452027180
Test: m --no-skip-soong-tests
Change-Id: I8306a7c11e3cb8079005c03a3ca5704b2d8b9715