Collect transitive deps of the host tools
This change creates soong module level dependency between modules and
the host tool modules that are used in RuleBuilderCommand. With this
change, Soong modules are now enforced to define the dependencies in the
DepsMutator to utilize the host tool in the RuleBuilderCommand. Note
that the Singletons do not have to specify this dependency.

Test: m nothing
Bug: 456238311
Change-Id: I9a249f705e88b2a0d9494369b705c752f157cb60