Use target.Variations in OtherModuleFarDependencyVariantExists search
filesystem uses OtherModuleFarDependencyVariantExists to conditionally
include a specific variation of a dependency created by `link` or
`rust_libraries` mutator. When variants are created on demand in the
future, this would require searching through the combinatorial of
(os,arch) mutations. To prune this search, make os and arch explicit in
the `OtherModuleFarDependencyVariantExists` call.

Bug: 448182248
Test: m nothing
Change-Id: If74d4b1346e6c33bc2a06a864fd24e8f27ae6739