Fix directory name for darwin.
The new Rust toolchains for darwin are in prebuilts/rust-toolchain/darwin,
not darwin-x86. This has necessitated a variety of workarounds, including this one, because many places compute directory names based on <os>-<arch>.

We decided to call it prebuilts/rust-toolchain/darwin in order to be forward-looking about migrating darwin host builds to arm64.

Test: m cargo_embargo (on Darwin)
Bug: 452035644
Change-Id: Ic985267c99adcbaff0072850c97face907cda510