r8 depfile verifier
proguard files can include other proguard files, which aren't listed
as dependencies currently. Add a check that proguard files are correctly
listed as dependencies by comparing the depfile produced by r8 against
a list of inputs.

This is essentially a "lite" form of action sandboxing, just for the
r8 rule.

The check is currently commented out, but in a followup I'll work on
properly specifying the dependencies and enabling it.

Rust will follow the same pattern, Ivan was working on something similar
in ag/34151397, but was asked to rewrite his tool in go, which I've done
here. I'll port his cl to this tool in a followup.

Bug: 436992217
Test: m SystemUI; It catches issues.
Change-Id: Ie9f06b2a167a2e743d65ebbc0dca11c5ba028921