[soong] Dist super_empty.img in dist_files goal
Currently, when BUILDING_SUPER_EMPTY_IMAGE is enabled, the
super_empty.img is built by Soong. However, the android_device.go file
within the filesystem package was not including super_empty.img in the
dist_files goal. This caused the super_empty.img to not be available
in the out/dist directory when running m dist.

This change updates the distFiles function in android_device.go to
explicitly check for the presence of super_empty.img via the
SuperImageInfo provider. If super_empty.img exists, it will now be
correctly added to the dist_files goal, making it available in out/dist
alongside super.img. This aligns the Soong build behavior with the
expected distribution of build artifacts.

Bug: 461986912

Bug: 452568527
Bug: 456000560
Bug: 455999422
Bug: 456000002
Test: m -j dist
Change-Id: I9798ff7094e51190ab0887f23afa328b255485d9