Remove PrebuiltGetPreferred
PrebuiltGetPreferred uses WalkDepsProxy, which pessimizes incremental
analysis.  It is only used to manually handle HostToolProviderInfo  of
prebuilts that replace sources in cases where the dependency was added
by a FinalDepsMutator, too late for the Prebuilts mutators to replace
the dependency.

Replace it with manually propagating the prebuilt's HostToolProviderInfo
(renamed to HostToolInfo) into the source module so that the users only
need to access their direct dependency.

Bug: 457275077
Test: no change to generated ninja files
Change-Id: I11642ba66f70fb4ae4eef8ea733adc186fd59fa4