Add included_proguard_flags_files
A proguard file can use "-include" to include another proguard file.
These files need to be known ahead of time for action sandboxing and
RBE. Add an included_proguard_flags_files property that these files
can be listed in, which will cause them to be added to the implicit
deps of the r8 rule.

ag/37062107 sets up a valdiation test to make sure all the included
files are listed properly. I'll work on adding the entries where
necessary and enabling the check in a followup.

Bug: 436992217
Test: Manually enable the validation, add missing files to SystemUI, and build.
Test: /google/src/head/depot/google3/experimental/users/colefaust/one_off_scripts/verify_ninja_files_unchanged.py main-copy main aosp_cf_x86_64_only_phone-trunk_staging-userdebug
Change-Id: I40a44c4e14edf1f9ff394795c4132cf85649b4f5