soong: check prebuilt boot image existence
If a product generates a prebuilt boot image during build (e.g. copying
from another location), it won't be available here for Soong to use.
Skip autogenerating the boot image in this case for Soong-only builds.

This is similar to http://ag/34108442 which skipped prebuilt DTBO images
that are generated during build.

Bug: b/459799623
Test: `m` now succeeds with http://ag/36999227
Change-Id: I86563f62130afe866f29e1b7b41470fa2c70fb69