Refactor: Use toybox cpio for SBOM testing
The SBOM integration tests currently rely on system's `cpio` command to
list the contents of ramdisk images. However, `cpio` is not guaranteed
to be installed by default in the build environment.

This change refactors the `test_sbom_aosp_cf_x86_64_phone` function to
use `toybox cpio` instead of the system's `cpio`. Moreover, this commit
introduces a checker which stops the test if any required tools are not
available.

Bug: 461395295
Test: presubmit
Change-Id: I06103c2d21014e415038ba40e3dae386d1c6fe61