Move rs global includes out of variable
Incremental analysis has to keep all variables, as they might be
referenced by cached ninja files and not visible to liveGlobals.
Renderscript uses pctx.PrefixedExistentPathsForSourcesVariable in
a global variable to generate the global include directory list
into the RsGlobalIncludes global variable.

The incremental analysis integration tests are not configured to
analyze any modules that have renderscript sources, so the full
build comparison strips the RsGlobalIncludes variable and doesn't
add dependencies on the list of include directories.  The incremental
analysis keeps the variable, resulting in a difference in the
globs results file.

Move the global include directories from a global variable into
the modules so they are consistently not present in the integration
tests.  There are only 10 modules that use renderscript sources,
so this won't significantly affect ninja file size or analysis
time.

Bug: 463538791
Test: build/soong/tests/build_action_caching_test.sh
Change-Id: I49bb81bbe6e3bf9ef8c3c8dcdae3ccfb847267d1