Output empty depfile when not using r8
The "incremental java compilation" rules sometimes use r8, in which
case a depfile is outputted, and sometimes only use d8, in which
case it isn't. The d8 case was failing because we are now expecting
the depfile as an output file.

Just create an empty output file in the d8 case.

Bug: 436992217
Test: lunch aosp_cf_x86_64_only_phone-trunk_staging-eng; SOONG_USE_PARTIAL_COMPILE=true SOONG_PARTIAL_COMPILE=all m out/soong/.intermediates/external/okhttp/okhttp/android_common_apex31/dex/okhttp.jar
Change-Id: Iac8b49a8dd687896d969c400b69b7c7497c9dba7