sbom_test: Fix ramdisk file listing for toybox cpio
The previous logic for listing files within `ramdisk.img` in `sbom_test.sh` is designed to parse the long listing output (similar to `ls -l`). This includes using `grep '^[-l]'` to filter for regular files and symlinks and awk command
to extract the filename.

However, the `cpio` command is aliased to `toybox cpio` in this script,
and the `toybox cpio -tv` does not produce a long listing but only outputs the filenames, one per line.

This commit introduces a method for obtaining the file list by
extracting the file in actual, and use find command to determine whether
it is a regular file or a symbolic link.

Bug: 461395295
Test: build/soong/tests/sbom_test.sh
Change-Id: I013992b64fa49fd0de67ccc7b3b6a9708c2757b0