droidstubs: always pass in flag information to metalava
We used to not pass in any flags to metalava when building the
"everything" set. However, metalava needs to know if flags are exported
or not in checks like "checkapi".

On the other hand, for the "everything" set, metalava considered the
omission of flag data as if all flags were enabled. Preserve that
behaviour by adding two new options to aconfig-to-metalava-flags to tell
the script to pretend that all flags have a given state or permission.
That includes all aconfig flags as well as any unknown flags. The latter
is handled by adding an `<unknown-flags>` element to the flag config
file.

Bug: 465655765
Test: m checkapi
Change-Id: I601f1a8a18ddac7192569a25d76ae822bfeb869c