Add GetModuleProxy to SingletonContext
Singletons can already get whatever module they want via
VisitAllModules() and checking the name, but this is obviously slow.
Add a method to directly get a module by name+variant.

This is intended for use in action sandboxing, to get host tools
in singletons. In regular modules we can add dependencies instead.

Bug: 448484009
Test: m nothing (switched a usage of VisitAllModuleProxies to GetModule)
Change-Id: If876ec63d67544be094bef379b4aa6c7da009669