Simplify bootclasspath profile configuration.
Previously, boot image profiles were imported using a ad-hoc array. This
CL changes it so that a module can declare its profile information using
a path instead.

This improves the correctness of boot image profiles by ensuring that
the profile is declared directly in the definition of the module that
provides the profile and that it is only added if the profile is
present.

Bug: 342163020
Test: `lunch cf_x86_64_phone-next-userdebug && m -j 64` shows a successful build using the existing pre-built modules
Test: Make sure that a profile is deapexed from the prebuilt non-art APEX (e.x., StatsD) https://paste.googleplex.com/6040870866518016
Test: Make sure the new `dex_preopt` field is generated only for current pre-builts:
[ART] https://paste.googleplex.com/6326743604527104,
[Statsd] https://paste.googleplex.com/6145511094616064
Test: go/art-benchmark?p=BootImageProfileId:58609 shows no change in terms of generated code file sizes
Test: go test ./... # Soong tests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I76244f7a913ab45b7601e7eea4d3926927d3ba5f