Rename toybox tools from NameTool to just Name
These tool definitions will be used inline in commands in RuleParams,
so it's kindof important that they're short.

Some rules (Cat, Cp, and Touch) were already using those names, so
those have been renamed to e.g. CatRule. The rule is usually used on
a line all on its own so it's not important for those to be short.

Bug: 448484009
Test: m nothing --no-skip-soong-tests
Change-Id: I8de6ef4792b2d2fce4fefa3bd861b08461f5af0d